+++ Changelog.rst
@@ -3,6 +3,12 @@ Changelog
 
 Summary of all changes made since the first stable release
 
+0.2.1 (11-24-2020)
+------------------
+* DOC: Updated examples in README
+* BUG: Fixed an error in determining the sign and direction of OCB vectors
+* STY: Changed a ValueError in VectorData to logger warning
+
 
 0.2.0 (10-08-2020)
 ------------------